Add files via upload
This project was quite an interesting (and challenging-at the latter parts) project. It sought to look into data from the "Movie Analytics" Database, containing more than 15 tables. Among the relevant information wanted was the top-5 movies in terms of budget, top and bottom 10 movies in terms of popularity among others. It was really intriguing as stated earlier, and recommended for all who want to test themselves.