Add files via upload
This project involves using the data to extract and gain insights into the root causes behind the spate of accidents within a certain country. Among other things, it wanted to know the frequency of accidents within the urban areas as compared to that of rural areas, whether the age of vehicles contributed in any way to the frequency of accidents etc.,  etc.